fix(syntax): align query files with grammar node types + wire mode activation to syntax rebuild
Query fixes for all 15 new language grammars — each query file
now references only node types and field names that exist in its
specific grammar version:

- Bash: removed until_statement, assignment_word (not in v0.23.3)
- CSS: keyframe_block_statement -> keyframes_statement
- Go: switch_statement -> expression/type_switch_statement
- HTML: removed id/class_attribute (not in v0.23.0)
- Java: switch_statement -> switch_expression; comment -> line/block_comment
- JSON: removed body: field on object
- Lua: variable_assignment -> assignment_statement; removed parameter
- Ruby: if/unless body: -> consequence:
- SQL: rewritten for tree-sitter-sequel node names
- TOML: rewritten for tree-sitter-toml-ng node names
- TypeScript: module_declaration -> internal_module; block -> statement_block;
  enum_declaration name type fix; interface_declaration body type fix

Plus:
- tree_resource.rs: migrate Node::child(i) from usize to u32 for ts 0.26 API
- lang.rs: detect shell dotfiles (.zshrc, .bashrc, .profile, etc.) as Bash
- modes.rs: add lang_for_mode_id reverse mapping
- dispatch.rs: rebuild syntax handle on major mode activation so
  manually toggling a language mode (:bash-mode) activates tree-sitter
  highlighting

+++ crates/lattice-plugin-host/src/tree_resource.rs
@@ -79,7 +79,7 @@ fn point_le(a: Point, b: Point) -> bool {
 fn resolve<'t>(tree: &'t Tree, path: &[u32]) -> Option<Node<'t>> {
     let mut node = tree.root_node();
     for &i in path {
-        node = node.child(i as usize)?;
+        node = node.child(i)?;
     }
     Some(node)
 }
@@ -94,7 +94,7 @@ fn path_of(node: Node) -> Vec<u32> {
     while let Some(parent) = cur.parent() {
         let mut idx = 0u32;
         for i in 0..parent.child_count() {
-            if parent.child(i).map(|c| c.id()) == Some(cur.id()) {
+            if parent.child(i as u32).map(|c| c.id()) == Some(cur.id()) {
                 idx = i as u32;
                 break;
             }
@@ -340,7 +340,7 @@ impl NodeResource {
         let node = resolve(tree, &self.path)?;
         let mut seen = 0u32;
         for i in 0..node.child_count() {
-            let child = node.child(i)?;
+            let child = node.child(i as u32)?;
             if child.is_named() {
                 if seen == index {
                     let mut path = self.path.clone();
@@ -384,11 +384,11 @@ impl NodeResource {
         let cur = last as usize;
         let candidate = if forward {
             (cur + 1..parent.child_count()).find(|&i| {
-                parent.child(i).map(|c| c.is_named()).unwrap_or(false)
+                parent.child(i as u32).map(|c| c.is_named()).unwrap_or(false)
             })
         } else {
             (0..cur).rev().find(|&i| {
-                parent.child(i).map(|c| c.is_named()).unwrap_or(false)
+                parent.child(i as u32).map(|c| c.is_named()).unwrap_or(false)
             })
         };
         candidate.map(|i| {
@@ -438,7 +438,7 @@ impl CursorResource {
             return false;
         };
         for i in 0..node.child_count() {
-            if node.child(i).map(|c| c.is_named()).unwrap_or(false) {
+            if node.child(i as u32).map(|c| c.is_named()).unwrap_or(false) {
                 self.path.push(i as u32);
                 return true;
             }
@@ -458,7 +458,7 @@ impl CursorResource {
             return false;
         };
         for i in (last as usize + 1)..parent.child_count() {
-            if parent.child(i).map(|c| c.is_named()).unwrap_or(false) {
+            if parent.child(i as u32).map(|c| c.is_named()).unwrap_or(false) {
                 let plen = self.path.len();
                 self.path[plen - 1] = i as u32;
                 return true;
